+++ lib/kitchen/directions/bake_autotitled_exercise/main.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Kitchen
+  module Directions
+    module BakeAutotitledExercise
+      def self.v1(exercise:, number: nil)
+        V1.new.bake(exercise: exercise, number: number)
+      end
+    end
+  end
+end