+++ spec/directions/bake_notes/bake_autotitled_notes_spec.rb
@@ -72,6 +72,24 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
     )
   end
 
+  let(:exercise_within_note) do
+    book_containing(html:
+      one_chapter_with_one_page_containing(
+        <<~HTML
+          <div data-type="note" id="untitlednote" class="foo">
+          <p>this is a note</p>
+          <div data-type="exercise" id="3360">
+            <div data-type="problem" id="504">
+              <ul>
+                <li>What do you need to know to perform this analysis at the very minimum?</li>
+              </ul>
+            </div>
+          </div>
+        HTML
+      )
+    )
+  end
+
   before do
     stub_locales({
       'iframe_link_text': 'Click to view content',
@@ -81,6 +99,9 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
         'project': 'Project',
         'media-2': 'Media',
         'interactive': 'Link to Learning'
+      },
+      'exercises': {
+        'exercise': 'Exercise'
       }
     })
   end
@@ -194,6 +215,33 @@ RSpec.describe Kitchen::Directions::BakeAutotitledNotes do
               </div>
             </li>
               </ul>
+            </div>
+          </div>
+        </div>
+      HTML
+    )
+  end
+
+  it 'bakes unclassified exercises within autotitled notes' do
+    described_class.v1(book: exercise_within_note, classes: %w[foo], bake_exercises: true)
+    expect(exercise_within_note.body.pages.first.notes).to match_normalized_html(
+      <<~HTML
+        <div class="foo" data-type="note" id="untitlednote">
+          <h3 class="os-title" data-type="title">
+            <span class="os-title-label">Bar</span>
+          </h3>
+          <div class="os-note-body">
+            <p>this is a note</p>
+            <div data-type="exercise" id="3360">
+              <div data-type="problem" id="504">
+                <span class="os-title-label">Exercise </span>
+                <span class="os-number">1</span>
+                <div class="os-problem-container">
+                  <ul>
+                    <li>What do you need to know to perform this analysis at the very minimum?</li>
+                  </ul>
+                </div>
+              </div>
             </div>
           </div>
         </div>